+++ backend/.gitignore
@@ -1,21 +1,75 @@
-# Node.js
-node_modules/
-npm-debug.log
-yarn-error.log
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*.so
 
-# Logs
-logs/
+# Virtual environment
+venv/
+ENV/
+env/
+
+# Distribution / packaging
+build/
+dist/
+*.egg-info/
+.eggs/
+
+# Installers
+*.exe
+*.dmg
+*.pkg
+*.deb
+
+# Logs and databases
 *.log
-*.log.*
+*.sql
+*.sqlite3
+*.db
 
-# Environment variables
+# Python environment and packages
+.Python
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+*.cover
+*.py,cover
+.cache
+nosetests.xml
+coverage.xml
+*.coveragerc
+
+# PyInstaller
+# Usually these files are written by a python script from a template
+# before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# dotenv
 .env
 
-# OS generated files
-.DS_Store
-Thumbs.db
+# Flask cache
+instance/
+*.mo
+*.pot
 
-# IDEs and editors
+# Mac OS-specific files
+.DS_Store
+
+# Jupyter Notebook
+.ipynb_checkpoints/
+
+# VSCode settings
 .vscode/
+
+# IDE / Editor-specific files
 .idea/
-*.swp
+*.sublime-workspace
+*.sublime-project
+
+# Docker
+*.dockerfile
