+++ backend/app/config.py
@@ -1,0 +1,5 @@
+import os
+
+
+class Config:
+    GOOGLE_API_KEY = os.getenv('GOOGLE_API_KEY')
