+++ backend/requirements.txt
@@ -1,0 +1,4 @@
+Flask
+Flask-Cors
+python-dotenv
+googlemaps
