+++ frontend/.env
@@ -1,0 +1,1 @@
+GOOGLE_API_KEY=AIzaSyA2TPXGn27OVkexuaK4iwcXMmAKzhrr09g
